flask_service/requirements.txt
-i https://pypi.org/simple
blinker==1.5
click==8.1.3 ; python_version >= '3.7'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
email-validator==1.3.1
flask==2.2.3
flask-cors==3.0.10
flask-mongoengine==1.0.0
flask-wtf==1.1.1 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
mongoengine==0.26.0 ; python_version >= '3.7'
pymongo==4.3.3 ; python_version >= '3.7'
python-dotenv==0.21.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
watchdog==2.2.1
werkzeug==2.2.3 ; python_version >= '3.7'
wtforms[email]==3.0.1 ; python_version >= '3.7'
